requirements.txt
weave
groq
instructor
e2b_code_interpreter
openai
python-dotenv